refactor(runner): name the serializer for what it produces, not for JSON
Groundwork, deliberately inert: nothing yet asks for the new member, and
no behavior changes. `scripts/ab-harness.ts` reports the encoded form,
the liveness scan and the provenance chain all IDENTICAL to `main`, on
real and synthetic patterns alike.

**`createJsonSchema` moves to its own file.** It infers a JSON Schema
from an example value, which is a different subject from everything else
in `json-utils.ts`, and it is the only member of that file exported
publicly (`runner/src/index.ts`, consumed by `packages/llm`). Its tests
move with it.

**The two serializers are named for what they answer.**
`moduleToJSON` -> `moduleToEncodableForm`, `patternToJSON` ->
`patternToEncodableForm`. Neither is used outside `packages/runner`, and
neither produces JSON: each answers a record the data model can
represent, which reaches storage without being stringified.

**A module gains `toEncodableForm`, and `toJSON` delegates to it.**
The new member is nascent -- the walk and the lookups still ask by
`toJSON`, so nothing reaches it yet. `toJSON` becomes one shared method
rather than a closure per module, taking its receiver as `this`, so it
goes directly into each module's literal.

Three constraints on that method are recorded in `json-member.ts`,
because none is visible from a call site: `this` is safe there and would
NOT be on `toEncodableForm` (which reads its receiver's own members, and
a factory carries `asScope` besides); it must be a function declaration
rather than a `const`, since a module is built during module evaluation
inside an import cycle where a `const` is still in its temporal dead
zone; and it stays in a leaf module so `builder/pattern.ts` can reach it
without closing a cycle.

**`moduleToEncodableForm` drops the serializer under BOTH names.** This
is what keeps the change inert: an extra member reaching `...rest` would
change the serialized key set, and with it every content-derived id for
every value carrying a module. `json-utils.test.ts` asserts the whole key
set, so it fails rather than drifting.

The A/B harness now asks for the pattern serializer under either name.
It runs the same probes against two checkouts, so a single-name lookup
would report a failure in whichever tree spells it the other way.
