+++ packages/runner/src/builder/module.ts
@@ -17,6 +17,7 @@ import type {
   SchemaWithoutCell,
   Stream,
   StripCell,
+  toEncodableForm,
   toJSON,
 } from "./types.ts";
 import { toCompactDebugString } from "@commonfabric/data-model/value-debug";
@@ -26,7 +27,8 @@ import {
   connectInputAndOutputs,
 } from "./node-utils.ts";
 import { assertNotInActionExecution } from "./action-context.ts";
-import { moduleToJSON } from "./json-utils.ts";
+import { moduleToEncodableForm } from "./json-utils.ts";
+import { toJSONMethod } from "./json-member.ts";
 import {
   brandTrustedBuilderArtifact,
   getArtifactEntryRef,
@@ -130,9 +132,10 @@ export function createNodeFactory<T = any, R = any>(
     moduleSpec.implementation = implementation;
   }
 
-  const module: Module & toJSON = {
+  const module: Module & toEncodableForm & toJSON = {
     ...moduleSpec,
-    toJSON: () => moduleToJSON(module),
+    toJSON: toJSONMethod,
+    toEncodableForm: () => moduleToEncodableForm(module),
   };
   // A module with ifc confidentiality on its argument schema should have at least
   // that value on its result schema
@@ -140,15 +143,18 @@ export function createNodeFactory<T = any, R = any>(
     module.argumentSchema,
     module.resultSchema,
   );
-  const factory = Object.assign((inputs: FactoryInput<T>): Reactive<R> => {
-    const outputs = reactive<R>(undefined, module.resultSchema);
-    const node: NodeRef = { module, inputs, outputs, frame: getTopFrame() };
+  const factory = Object.assign(
+    (inputs: FactoryInput<T>): Reactive<R> => {
+      const outputs = reactive<R>(undefined, module.resultSchema);
+      const node: NodeRef = { module, inputs, outputs, frame: getTopFrame() };
 
-    connectInputAndOutputs(node);
-    (outputs as OpaqueCell<R>).connect(node);
+      connectInputAndOutputs(node);
+      (outputs as OpaqueCell<R>).connect(node);
 
-    return outputs;
-  }, module) as ModuleFactory<T, R>;
+      return outputs;
+    },
+    module,
+  ) as ModuleFactory<T, R>;
   factory.asScope = (scope: CellScope) =>
     createNodeFactory({ ...module, defaultScope: scope });
   // Provenance brand: every node factory (lift / handler / byRef / the list-op
@@ -428,7 +434,7 @@ function handlerInternal<E, T>(
     stateSchema as JSONSchema | undefined,
   );
 
-  const module: Handler<E, T> & toJSON & {
+  const module: Handler<E, T> & toEncodableForm & toJSON & {
     bind: (inputs: FactoryInput<StripCell<T>>) => Stream<E>;
   } = {
     type: "javascript",
@@ -438,7 +444,8 @@ function handlerInternal<E, T>(
     // Overriding the default `bind` method on functions. The wrapper will bind
     // the actual inputs, so they'll be available as `this`
     bind: (inputs: FactoryInput<StripCell<T>>) => factory(inputs),
-    toJSON: () => moduleToJSON(module),
+    toJSON: toJSONMethod,
+    toEncodableForm: () => moduleToEncodableForm(module),
     ...(schema !== undefined && { argumentSchema: schema }),
     ...(writableProxy && { writableProxy: true }),
   };
