+++ packages/runner/test/content-addressed-identity-adversarial.test.ts
@@ -786,7 +786,7 @@ describe("content-addressed identity — adversarial (C5 red-team gate)", () => 
     });
 
     it("a dynamic artifact (no symbol) serializes without $implRef", async () => {
-      // moduleToJSON only emits $implRef when provenance.symbol is present, so a
+      // moduleToEncodableForm only emits $implRef when provenance.symbol is present, so a
       // dynamic (symbol-less) artifact never gets a serialized, reload-resolvable
       // reference — confirming in-session-only authority on the serialization seam.
       const pattern = await setup();
@@ -809,10 +809,14 @@ describe("content-addressed identity — adversarial (C5 red-team gate)", () => 
         implementationRef: "dyn-ref",
         toJSON: undefined as unknown,
       };
-      // moduleToJSON is reached via the builder; call the same path the real
+      // moduleToEncodableForm is reached via the builder; call the same path the real
       // module uses. We re-import it lazily to avoid widening the import surface.
-      const { moduleToJSON } = await import("../src/builder/json-utils.ts");
-      const json = moduleToJSON(dynModule as unknown as Module) as Record<
+      const { moduleToEncodableForm } = await import(
+        "../src/builder/json-utils.ts"
+      );
+      const json = moduleToEncodableForm(
+        dynModule as unknown as Module,
+      ) as Record<
         string,
         unknown
       >;
